Container folder is moved out of Components
