+++ src/client/container/CommentBox/blogCard.css
@@ -1,103 +1,129 @@
 /*class for div whole container */
-.work-panel{
+
+.work-panel {
     position: absolute;
     width: 498px;
     height: 239px;
     left: 222px;
     top: 107px;
 }
-.blog-card{   
+
+.blog-card {
     width: 280px;
-    height: 90px;  
+    height: 90px;
     display: flex;
-    flex-direction: column;    
-    justify-content: space-evenly;  
-    margin-left: 10px; 
+    flex-direction: column;
+    justify-content: space-evenly;
+    margin-left: 10px;
     border-style: solid;
     border-width: thick;
     border-color: hsl(71, 53%, 77%);
     margin-bottom: 20px;
+    position: fixed;
 }
+
+
 /*class for div consists of the lable and input */
-.inputBox{
-    width: 50%;
+
+.inputBox {
+    width: 100%;
     margin-left: 10px;
     display: flex;
     flex-direction: column;
-    justify-content: center; 
-    align-content: center;     
-    position: absolute;
-    z-index: 2;
+    justify-content: center;
+    align-content: center;
+    position: relative;
+    z-index: auto;
 }
-.blog-card-label{
+
+.blog-card-label {
     font-family: serif;
     font-size: 12px;
-    margin-bottom: 16px;
+    margin-bottom: 50px;
     margin-top: 0px;
 }
+
+
 /*class for input */
-.blog-card-control{
+
+.blog-card-control {
     width: 240px;
     height: 18px;
     border-radius: 10px;
     background-color: whitesmoke;
-    font-family: sans-serif;  
-    font-size: 10px;    
+    font-family: sans-serif;
+    font-size: 10px;
 }
+
+
 /* class for div of consists of dot button list*/
-.btn-right{ 
+
+.btn-right {
     font-size: 16px;
     width: 30px;
     height: 100px;
     margin-bottom: 2px;
-    margin-top: 2px; 
-    text-align: center;       
+    margin-top: 2px;
+    text-align: center;
 }
-button.btn-right{
+
+button.btn-right {
     width: 0px;
     height: 0px;
-    border:none;
-    background-color:white;
-    outline:none;
+    border: none;
+    background-color: white;
+    outline: none;
     cursor: pointer;
-    margin-top: 6px;
+    margin-top: 40px;
     margin-right: 20px;
 }
+
+
 /*class for div consists .btn-right and .dropdown-contentcontent*/
-.menu-right{    
+
+.menu-right {
     display: flex;
+    flex-direction: row;
     justify-content: flex-end;
 }
+
+
 /* display property for .dropdown-content */
+
 .dropdown-content {
     display: none;
     flex-direction: column;
     justify-content: flex-start;
-    position: absolute;
-    margin-top: 15px;    
+    margin-top: 500px;
     width: 80px;
     height: 100px;
-    overflow: auto;    
-    z-index: -3;    
+    overflow: auto;
+    z-index: -3;
 }
+
+
 /*class for for each list items*/
-.listComponent{
+
+.listComponent {
     list-style-type: none;
     cursor: pointer;
     font-size: 12px;
 }
-.listComponent:hover{
-    background-color:  #f1f1f1;
+
+.listComponent:hover {
+    background-color: #f1f1f1;
 }
-.btnComponent{
-    border:none;
-    background-color:whitesmoke;
-    outline:none;
+
+.btnComponent {
+    border: none;
+    background-color: whitesmoke;
+    outline: none;
     cursor: pointer;
     font-size: 10px;
 }
-.comment-list{  
-    position: relative;  
+
+.comment-list {
+    position: relative;
     margin-top: 20px;
     margin-bottom: 20px;
     margin-left: 20px;
