+++ src/client/container/CommentList/CommentList.component.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import List from "../../List/List.component";
+import List from "../../components/List/List.component";
 
 const CommentList = props => {
