+++ src/client/container/DropDown/DropDown.component.jsx
@@ -10,9 +10,5 @@ const DropDown = props => {
     </div> 
   )     
 }
-/*const dropdownStyle = {
-  display: 'flex',
-  zIndex: '4',
-  marginTop: '25px'
-}*/
+
 export default DropDown;
